requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
PyMySQL==1.1.1
SQLAlchemy==2.0.34
Werkzeug==3.0.4
python-dateutil==2.9.0.post0
